src/requirements.txt
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.19.1
python-dotenv==1.0.0
pyarrow==11.0.0
